handleSubmitAmountをカスタムフックに切り出し
