+++ frontend/src/components/pages/NewPaymentEntry.tsx
@@ -1,25 +1,17 @@
-import { useContext, VFC } from 'react'
+import { VFC } from 'react'
 import { useForm } from 'react-hook-form'
-import { useHistory } from 'react-router-dom'
 
 import { Flex, FormControl, FormErrorMessage, FormLabel, Grid, Input } from '@chakra-ui/react'
-import axios from 'axios'
 import { DateTime } from 'luxon'
 
 import { PrimaryButton } from 'components/atoms/button/PrimaryButton'
 import { HeaderWithTitleLayout } from 'components/templates/HeaderWithTitleLayout'
-import { PaymentContext } from 'context/PaymentContext'
-import { postPayment } from 'lib/api/payment'
-import { auth } from 'lib/firebase'
-import { useToast } from 'lib/toast'
-import { MultipleErrorResponse } from 'types/multipleErrorResponses'
+import { useSubmitAmount } from 'hooks/useSubmitAmount'
 
 import type { PostPaymentParams } from 'types/postPaymentParams'
 
 export const NewPaymentEntry: VFC = () => {
-  const { updatePaymentList, setUpdatePaymentList } = useContext(PaymentContext)
-  const { errorToast, successToast } = useToast()
-  const history = useHistory()
+  const { handleSubmitAmount } = useSubmitAmount()
   const {
     register,
     handleSubmit,
@@ -31,24 +23,6 @@ export const NewPaymentEntry: VFC = () => {
       paidAt: DateTime.local().toFormat('yyyy-MM-dd'),
     },
   })
-
-  const handleSubmitAmount = async (params: PostPaymentParams) => {
-    const idToken = await auth.currentUser?.getIdToken(true)
-    try {
-      await postPayment(params, idToken)
-      setUpdatePaymentList(!updatePaymentList)
-      history.push('/')
-      successToast('支払い情報を登録しました')
-    } catch (err) {
-      if (axios.isAxiosError(err) && err.response) {
-        ;(err.response?.data as MultipleErrorResponse).messages.forEach((message) => {
-          errorToast(message)
-        })
-      } else {
-        errorToast('エラーが発生しました')
-      }
-    }
-  }
 
   return (
     <HeaderWithTitleLayout title="支払い情報の入力">
@@ -83,7 +57,8 @@ export const NewPaymentEntry: VFC = () => {
                 id="detail"
                 type="text"
                 size="lg"
-                placeholder="例）スーパー" // eslint-disable-next-line react/jsx-props-no-spreading
+                placeholder="例）スーパー"
+                // eslint-disable-next-line react/jsx-props-no-spreading
                 {...register('detail', {
                   maxLength: {
                     value: 28,
@@ -98,7 +73,8 @@ export const NewPaymentEntry: VFC = () => {
               <Input
                 id="paid_at"
                 type="date"
-                size="lg" // eslint-disable-next-line react/jsx-props-no-spreading
+                size="lg"
+                // eslint-disable-next-line react/jsx-props-no-spreading
                 {...register('paidAt', {
                   required: '支払日を入力してください',
                 })}
